pyproject.toml
[project]
name = "air.traffic.trends-CO"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "", email = ""},
]
dependencies = [
    "requests>=2.31.0",
    "beautifulsoup4>=4.12.3",
    "ipykernel>=6.29.4",
    "tqdm>=4.66.4",
    "pandas>=2.2.2",
    "polars>=0.20.23",
    "openpyxl>=3.1.2",
    "pyarrow>=16.0.0",
    "pydantic>=2.7.1",
    "numpy>=1.26.4",
    "cargo>=0.3",
    "boto3>=1.34.98",
    "pymysql>=1.1.0",
    "mysql-connector-python>=8.4.0",
    "psycopg2>=2.9.9",
    "python-dotenv>=1.0.1",
]
requires-python = "==3.12.*"
readme = "README.md"
license = {text = "MIT"}


[tool.pdm]
distribution = false
